Fix old compound v3 on several chains + add base
