+++ .vscode/launch.json
@@ -97,6 +97,7 @@
         "badger-dao",
         "balancer",
         "bancor-v3",
+        "baseswap",
         "beefy",
         "bella-protocol",
         "belt-finance",
