+++ src/adapters/compound-v3/arbitrum/index.ts
@@ -1,12 +1,9 @@
+import { getAssetsContracts } from '@adapters/compound-v3/common/asset'
+import { getCompLendBalances, getCompRewardBalances } from '@adapters/compound-v3/common/balance'
 import type { BalancesContext, BaseContext, Contract, GetBalancesHandler } from '@lib/adapter'
 import { resolveBalances } from '@lib/balance'
-import type { BalanceWithExtraProps } from '@lib/compound/v2/lending'
-import { getHealthFactor } from '@lib/compound/v2/lending'
+import { getSingleStakeBalances } from '@lib/stake'
 import type { Token } from '@lib/token'
-
-import { getAssetsContracts, getLendBorrowBalances } from '../common/lend'
-import { getRewardBalances } from '../common/rewards'
-import { getStakeBalances } from '../common/stake'
 
 const USDC: Token = {
   chain: 'arbitrum',
@@ -15,38 +12,36 @@ const USDC: Token = {
   symbol: 'USDC',
 }
 
-const CompoundUSDCv3: Contract = {
+const cUSDCv3: Contract = {
   chain: 'arbitrum',
   address: '0xA5EDBDD9646f8dFF606d7448e414884C7d905dCA',
   underlyings: [USDC],
 }
 
-const CompoundRewards: Contract = {
+const rewarder: Contract = {
   chain: 'arbitrum',
   address: '0x88730d254A2f7e6AC8388c3198aFd694bA9f7fae',
 }
 
 export const getContracts = async (ctx: BaseContext) => {
-  const assets = await getAssetsContracts(ctx, [CompoundUSDCv3])
+  const assets = await getAssetsContracts(ctx, [cUSDCv3])
 
   return {
-    contracts: { compounders: [CompoundUSDCv3], assets, CompoundRewards },
+    contracts: { compounders: [cUSDCv3], assets, rewarder },
   }
 }
 
 const compoundBalances = async (ctx: BalancesContext, compounders: Contract[], rewarder: Contract) => {
-  return Promise.all([getStakeBalances(ctx, compounders), getRewardBalances(ctx, rewarder, compounders)])
+  return Promise.all([getSingleStakeBalances(ctx, compounders), getCompRewardBalances(ctx, rewarder, compounders)])
 }
 
 export const getBalances: GetBalancesHandler<typeof getContracts> = async (ctx, contracts) => {
   const balances = await resolveBalances<typeof getContracts>(ctx, contracts, {
-    assets: (...args) => getLendBorrowBalances(...args, [CompoundUSDCv3]),
-    compounders: (...args) => compoundBalances(...args, CompoundRewards),
+    assets: (...args) => getCompLendBalances(...args, [cUSDCv3]),
+    compounders: (...args) => compoundBalances(...args, rewarder),
   })
 
-  const healthFactor = await getHealthFactor(balances as BalanceWithExtraProps[])
-
   return {
-    groups: [{ balances, healthFactor }],
+    groups: [{ balances }],
   }
 }
