+++ src/adapters/compound-v3/index.ts
@@ -1,6 +1,7 @@
 import type { Adapter } from '@lib/adapter'
 
 import * as arbitrum from './arbitrum'
+import * as base from './base'
 import * as ethereum from './ethereum'
 import * as polygon from './polygon'
 
@@ -9,6 +10,7 @@ const adapter: Adapter = {
   ethereum,
   polygon,
   arbitrum,
+  base,
 }
 
 export default adapter
