+++ src/adapters/index.ts
@@ -26,6 +26,7 @@ import babylonFinance from '@adapters/babylon-finance'
 import badgerDao from '@adapters/badger-dao'
 import balancer from '@adapters/balancer'
 import bancorV3 from '@adapters/bancor-v3'
+import baseswap from '@adapters/baseswap'
 import beefy from '@adapters/beefy'
 import bellaProtocol from '@adapters/bella-protocol'
 import beltFinance from '@adapters/belt-finance'
@@ -262,6 +263,7 @@ export const adapters: Adapter[] = [
   badgerDao,
   balancer,
   bancorV3,
+  baseswap,
   beefy,
   bellaProtocol,
   beltFinance,
